fix mox output not rising with heat

+++ src/main/java/gregtech/api/items/GT_RadioactiveCellIC_Item.java
@@ -84,16 +84,6 @@ public class GT_RadioactiveCellIC_Item extends GT_RadioactiveCell_Item implement
             }
         }
         if (getDamageOfStack(yourStack) >= getMaxDamageEx() - 1) {
-//            switch (this.numberOfCells) {
-//                case 1:
-//                    reactor.setItemAt(x, y, ItemList.Depleted_Thorium_1.get(1, new Object[0]));
-//                    break;
-//                case 2:
-//                    reactor.setItemAt(x, y, ItemList.Depleted_Thorium_2.get(1, new Object[0]));
-//                    break;
-//                case 4:
-//                    reactor.setItemAt(x, y, ItemList.Depleted_Thorium_4.get(1, new Object[0]));
-//            }
         	reactor.setItemAt(x, y, sDepleted.copy());
         } else if (heatrun) {
             damageItemStack(yourStack, 1);
@@ -121,9 +111,9 @@ public class GT_RadioactiveCellIC_Item extends GT_RadioactiveCell_Item implement
     }
 
     public boolean acceptUraniumPulse(IReactor reactor, ItemStack yourStack, ItemStack pulsingStack, int youX, int youY, int pulseX, int pulseY, boolean heatrun) {
-        if (!heatrun) {
+    	if (!heatrun) {
         	if(sMox){
-        	      float breedereffectiveness = reactor.getHeat() / reactor.getMaxHeat();
+        	      float breedereffectiveness = (float)reactor.getHeat() / (float)reactor.getMaxHeat();
         	      float ReaktorOutput = 1.5F * breedereffectiveness + 1.0F;
         	      reactor.addOutput(ReaktorOutput * this.sEnergy);
         	}else{
